+++ src/components/SearchDictionaryForm/SearchDictionaryFormContainer.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import SearchInputPresentation from 'components/SearchInput/SearchInputPresentation'
 import useSearchBoxNavigation from 'common/search/useSearchBoxNavigation'
 
-function SearchDictionaryContainer({ kids, customBaseUrl, searchType }) {
+function SearchDictionaryFormContainer({ kids, customBaseUrl, searchType }) {
   // basic searchbox behaviour
   const {
     handleSearchNavigation,
@@ -35,15 +35,15 @@ function SearchDictionaryContainer({ kids, customBaseUrl, searchType }) {
   }
 
   return (
-    <div id="DictionarySearchContainer" className="flex w-full rounded-lg">
+    <div id="SearchDictionaryFormContainer" className="flex w-full rounded-lg">
       <SearchInputPresentation
-        handleSearchSubmit={handleSearchNavigation}
-        handleTextFieldChange={handleSearchTermChange}
-        handleSearchLanguageChange={handleSearchLanguageNavigation}
+        handleSearchNavigation={handleSearchNavigation}
+        handleSearchTermChange={handleSearchTermChange}
+        handleSearchLanguageNavigation={handleSearchLanguageNavigation}
         searchLanguage={searchLanguage}
         searchLanguageOptions={searchLanguageOptions}
-        searchValue={displayedSearchTerm}
-        placeholder={searchBoxPlaceholder}
+        displayedSearchTerm={displayedSearchTerm}
+        searchBoxPlaceholder={searchBoxPlaceholder}
       />
 
       {submittedSearchTerm && (
@@ -61,10 +61,10 @@ function SearchDictionaryContainer({ kids, customBaseUrl, searchType }) {
 
 // PROPTYPES
 const { bool, string } = PropTypes
-SearchDictionaryContainer.propTypes = {
+SearchDictionaryFormContainer.propTypes = {
   customBaseUrl: string,
   kids: bool,
   searchType: string,
 }
 
-export default SearchDictionaryContainer
+export default SearchDictionaryFormContainer
