+++ src/components/SearchInput/SearchInputPresentation.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 
 // FPCC
@@ -7,52 +7,75 @@ import SearchLanguageSelector from 'components/SearchLanguageSelector'
 import SearchBox from 'components/SearchBox'
 
 function SearchInputPresentation({
+  handleSearchNavigation,
+  handleSearchLanguageNavigation,
+  handleSearchTermChange,
+  searchBoxPlaceholder,
   searchLanguage,
-  handleSearchSubmit,
-  handleTextFieldChange,
-  handleSearchLanguageChange,
   searchLanguageOptions,
-  placeholder,
-  searchValue,
+  displayedSearchTerm,
+  minimal,
 }) {
-  return (
-    <>
+  return minimal ? (
+    <div id="SearchInputPresentation-Minimal">
+      <div className="w-full flex rounded-full h-10">
+        <SearchBox.Presentation
+          onSubmit={handleSearchNavigation}
+          setSearchTerm={handleSearchTermChange}
+          searchTerm={displayedSearchTerm}
+          placeholder="Search"
+          minimal
+        />
+        <button
+          type="button"
+          id="SearchSubmit"
+          aria-label="Search/Go"
+          onClick={handleSearchNavigation}
+          className="relative inline-flex items-center px-2 py-1.5 text-fv-charcoal-light rounded-r-full bg-gray-50 hover:bg-gray-100"
+        >
+          {getIcon('Search', 'fill-current h-5 w-5 ')}
+        </button>
+      </div>
+    </div>
+  ) : (
+    <div id="SearchInputPresentation" className="flex w-full rounded-lg">
       <SearchBox.Presentation
-        onSubmit={handleSearchSubmit}
-        setSearchTerm={handleTextFieldChange}
-        searchTerm={searchValue}
-        placeholder={placeholder}
+        onSubmit={handleSearchNavigation}
+        setSearchTerm={handleSearchTermChange}
+        searchTerm={displayedSearchTerm}
+        placeholder={searchBoxPlaceholder}
       />
 
       <div className="relative inline-flex items-center px-2 py-1.5 text-fv-charcoal-light border-l-2 border-gray-300 rounded-r-md bg-gray-50 hover:bg-gray-100">
         <SearchLanguageSelector.Presentation
           selected={searchLanguage}
           options={searchLanguageOptions}
-          onSelect={handleSearchLanguageChange}
+          onSelect={handleSearchLanguageNavigation}
         />
         <button
           type="button"
           id="SearchSubmit"
           aria-label="Search/Go"
-          onClick={handleSearchSubmit}
+          onClick={handleSearchNavigation}
         >
           {getIcon('Search', 'fill-current h-7 w-7 ')}
         </button>
       </div>
-    </>
+    </div>
   )
 }
 
 // PROPTYPES
-const { func, object, string } = PropTypes
+const { bool, func, object, string } = PropTypes
 SearchInputPresentation.propTypes = {
+  handleSearchNavigation: func,
+  handleSearchLanguageNavigation: func,
+  handleSearchTermChange: func,
+  searchBoxPlaceholder: string,
   searchLanguage: string,
-  handleSearchSubmit: func,
-  handleTextFieldChange: func,
-  handleSearchLanguageChange: func,
   searchLanguageOptions: object,
-  placeholder: string,
-  searchValue: string,
+  displayedSearchTerm: string,
+  minimal: bool,
 }
 
 export default SearchInputPresentation
