+++ src/components/SearchInput/index.js
@@ -1,9 +1,5 @@
-import SearchInputContainer from 'components/SearchInput/SearchInputContainer'
-import SearchInputContainerMinimal from 'components/SearchInput/SearchInputContainerMinimal'
 import SearchInputPresentation from 'components/SearchInput/SearchInputPresentation'
 
 export default {
-  Container: SearchInputContainer,
-  ContainerMinimal: SearchInputContainerMinimal,
   Presentation: SearchInputPresentation,
 }
